Editing installationScript
Editing the installation script, by creating separate scripts for installing apps like brave, vscode, pycharm, and then linking them to the main script.

+++ scripts/apps/google-chrome.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 #
 
-sudo apt install wget -y
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo apt install ./google-chrome-stable_current_amd64.deb -y
-rm google-chrome-stable_current_amd64.deb
+if ! command -v google-chrome >/dev/null 2>&1; then
+    echo "** Installing Google Chrome...**"
+    sudo apt install wget -y
+    wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+    sudo apt install ./google-chrome-stable_current_amd64.deb -y
+    rm google-chrome-stable_current_amd64.deb
+else
+    echo "Google Chrome is already installed."
+fi
